+++ app/serializers/encryptable/file_serializer.rb
@@ -1,11 +1,11 @@
 # frozen_string_literal: true
 
 class Encryptable::FileSerializer < ApplicationSerializer
-  attributes :id, :name, :description, :sender_name
+  attributes :id, :name, :description, :sender_name, :created_at
 
   def sender_name
     object.sender&.label
   end
 
-  belongs_to :encryptable_credential
+  belongs_to :encryptable_credential, if: -> { object.encryptable_credential.present? }
 end