+++ frontend/app/models/encryptable-file.js
@@ -3,11 +3,9 @@ import { attr, belongsTo } from "@ember-data/model";
 
 export default class EncryptableFile extends Encryptable {
   @attr file;
-  @belongsTo("encryptable-credential", {
-    async: false,
-    inverse: "encryptableFiles"
-  })
+  @belongsTo("encryptable-credential")
   encryptableCredential;
+  @belongsTo("folder") folder;
 
   async save() {
     if (this.isDeleted) {
@@ -15,11 +13,13 @@ export default class EncryptableFile extends Encryptable {
     }
     const url = `/api/encryptables`;
     const credentialId = await this.encryptableCredential.get("id");
+    const folderId = await this.folder.get("id");
 
     const opts = {
       data: {
         description: this.description || "",
-        credential_id: credentialId
+        ...(credentialId !== undefined && { credential_id: credentialId }),
+        ...(folderId !== undefined && { folder_id: folderId })
       },
       headers: {
         "X-CSRF-Token": this.csrfToken