+++ frontend/tests/integration/components/encryptable/row-test.js
@@ -5,6 +5,7 @@ import { hbs } from "ember-cli-htmlbars";
 import Service from "@ember/service";
 import { isPresent } from "@ember/utils";
 import { setLocale } from "ember-intl/test-support";
+import EmberObject from "@ember/object";
 
 const storeStub = Service.extend({
   query(modelName, params) {
@@ -254,5 +255,42 @@ module("Integration | Component | encryptable/row", function (hooks) {
     assert.ok(isPresent(keyIcon));
     assert.ok(isPresent(deleteButton));
     assert.ok(isPresent(personIcon));
+  });
+
+  test("it renders with file", async function (assert) {
+    const folder = {
+      id: 1,
+      name: "Inbox",
+      isInboxFolder: true,
+
+      get(key) {
+        if (key === "isInboxFolder") {
+          return true;
+        }
+      }
+    };
+
+    this.set(
+      "encryptable",
+      EmberObject.create({
+        id: 1,
+        type: "encryptable_files",
+        name: "FolderFile",
+        description:
+          "This encryptable is attached to a folder, this is amazing. WOW!",
+        isFile: true,
+        folder: folder
+      })
+    );
+
+    await render(hbs`<Encryptable::Row @encryptable={{this.encryptable}}/>`);
+
+    await new Promise((r) => setTimeout(r, 10));
+
+    assert.equal(
+      this.element.querySelector("#encryptable-row-title").innerText,
+      "FolderFile"
+    );
+    assert.dom(this.element.querySelector("#download-button")).exists();
   });
 });